extract schema and resolvers for post example into separate files
