+++ src/server/api/graphqls/schema_def.graphqls
@@ -1,0 +1,29 @@
+# Database counter
+type Count {
+  # Current amount
+  amount: Int!
+}
+
+type Query {
+  # Counter
+  count: Count
+}
+
+type Mutation {
+  # Increase counter value, returns current counter amount
+  addCount(
+    # Amount to add to counter
+    amount: Int!
+  ): Count
+}
+
+type Subscription {
+  # Subscription fired when anyone increases counter
+  countUpdated(ids: String): Count
+}
+
+schema {
+  query: Query
+  mutation: Mutation
+  subscription: Subscription
+}
